move validator to utils
